block: move CompressionStats code
